同じプラグインの重複インストール時に警告を出し、上書きインストールを選択できるようにした

+++ plugin_manager.py
@@ -8,7 +8,7 @@ import shutil
 import json
 import importlib
 import flet as ft
-from typing import Callable
+from typing import Callable, Dict
 from interfaces.plugin_interface import PluginInterface
 from interfaces.system_plugin_interface import SystemPluginInterface
 from ui_component_manager import UIComponentManager
@@ -33,10 +33,21 @@ class PluginManager:
         self.plugin_folder_path = os.path.join(save_dir, PLUGIN_FOLDER)
         self.system_plugin_folder_path = os.path.join(base_dir, SYSTEM_PLUGIN_FOLDER)
         self.myapp_container = None
+        self.installed_plugins: Dict[str, str] = {}  # 追加: インストール済みプラグインの辞書 {plugin_name: plugin_dir}
+        self.load_installed_plugins_info()
         if not os.path.exists(self.plugin_folder_path):
             os.makedirs(self.plugin_folder_path)
         if not os.path.exists(self.system_plugin_folder_path):
             os.makedirs(self.system_plugin_folder_path)
+
+    def load_installed_plugins_info(self):
+        target_list = [filename for filename in os.listdir(self.plugin_folder_path) if not filename.startswith('.')]
+        for plugin_name in target_list:
+            plugin_dir = os.path.join(self.plugin_folder_path, plugin_name)
+            if os.path.isdir(plugin_dir):
+                with open(os.path.join(plugin_dir, "plugin.json"), 'r', encoding='utf-8') as f:
+                    plugin_info = json.load(f)
+                self.installed_plugins[plugin_info["plugin_name"]] = plugin_dir
 
     def _load_plugin(self, plugin_dir: str, container: ft.Container):
         try:
@@ -76,15 +87,31 @@ class PluginManager:
         if "system" in plugin_dir:
             container.controls.append(app_container_widget)
         else:
-            unique_key = str(uuid.uuid4())
-            deletable_app_container = ft.GestureDetector(
-                content=app_container_widget,
-                on_long_press_start=lambda e: self.show_delete_confirmation(plugin_dir, unique_key)
-            )
-            container.controls.append(deletable_app_container)
+            plugin_name = plugin_info["plugin_name"]
+            
+            if plugin_name in self.plugin_dict:
+                # 既存のプラグインアイコンを更新
+                self.plugin_dict[plugin_name].content = app_container_widget
+            else:
+                # 新しいプラグインアイコンを追加
+                deletable_app_container = ft.GestureDetector(
+                    content=app_container_widget,
+                    on_long_press_start=lambda e: self.show_delete_confirmation(plugin_dir, plugin_name)
+                )
+                container.controls.append(deletable_app_container)
+                self.plugin_dict[plugin_name] = deletable_app_container
+            
             self.myapp_container = container
-            self.plugin_dict[unique_key] = deletable_app_container
+            self.installed_plugins[plugin_name] = plugin_dir
         self.page.update()
+
+    def get_icon_base64(self, plugin_dir: str) -> str:
+        with open(os.path.join(plugin_dir, "plugin.json"), 'r', encoding='utf-8') as f:
+            plugin_info = json.load(f)
+        icon_path = os.path.join(plugin_dir, plugin_info["icon"])
+        with open(icon_path, "rb") as image_file:
+            encoded_string = base64.b64encode(image_file.read()).decode("utf-8")
+        return encoded_string
 
     def install_plugin(self, e: ft.FilePickerResultEvent, container: ft.Container) -> None:
         # プラグインを保存する一意のディレクトリを作成
@@ -101,13 +128,60 @@ class PluginManager:
                 shutil.copy(picked_file_path, zip_path)
                 with zipfile.ZipFile(zip_path, 'r') as zip_ref:
                     zip_ref.extractall(plugin_dir)
-            # セキュリティスキャンを実行
-            if not CodeSecurityScanner.scan_for_forbidden_functions(plugin_dir):
-                 self.api.logger.error("Security check failed: Forbidden functions found in the plugin code.")
-                 raise Exception("Security check failed: Forbidden functions found in the plugin code.")
-            self.api.logger.info(f"Security check passed for plugin: {plugin_dir}")
-            self._load_plugin(plugin_dir, container)
+                with open(os.path.join(plugin_dir, "plugin.json"), 'r', encoding='utf-8') as f:
+                    plugin_info = json.load(f)
 
+                plugin_name = plugin_info["plugin_name"]
+
+                if plugin_name in self.installed_plugins:
+                    def overwrite_plugin(choice):
+                        if choice == "yes":
+                            old_plugin_dir = self.installed_plugins[plugin_name]
+                            
+                            # 古いプラグインのディレクトリ内のファイルを削除
+                            for filename in os.listdir(old_plugin_dir):
+                                file_path = os.path.join(old_plugin_dir, filename)
+                                if os.path.isfile(file_path) or os.path.islink(file_path):
+                                    os.unlink(file_path)
+                                elif os.path.isdir(file_path):
+                                    shutil.rmtree(file_path)
+                            
+                            # 新しいプラグインのファイルを古いディレクトリにコピー
+                            for filename in os.listdir(plugin_dir):
+                                src_path = os.path.join(plugin_dir, filename)
+                                dst_path = os.path.join(old_plugin_dir, filename)
+                                if os.path.isfile(src_path) or os.path.islink(src_path):
+                                    shutil.copy2(src_path, dst_path)
+                                elif os.path.isdir(src_path):
+                                    shutil.copytree(src_path, dst_path)
+                            
+                            shutil.rmtree(plugin_dir)
+                            self._load_plugin(old_plugin_dir, container)
+                        else:
+                            shutil.rmtree(plugin_dir)
+                        self.page.dialog.open = False
+                        self.page.update()
+
+                    confirm_dialog = ft.AlertDialog(
+                        title=ft.Text("上書きインストールの確認"),
+                        content=ft.Text(f"同じ名前のプラグイン '{plugin_name}' が既にインストールされています。上書きインストールしますか？"),
+                        actions=[
+                            ft.TextButton("はい", on_click=lambda _: overwrite_plugin("yes")),
+                            ft.TextButton("いいえ", on_click=lambda _: overwrite_plugin("no")),
+                        ],
+                        actions_alignment=ft.MainAxisAlignment.END,
+                    )
+                    self.page.dialog = confirm_dialog
+                    self.page.dialog.open = True
+                    self.page.update()
+                else:
+                    # セキュリティスキャンを実行
+                    if not CodeSecurityScanner.scan_for_forbidden_functions(plugin_dir):
+                        self.api.logger.error("Security check failed: Forbidden functions found in the plugin code.")
+                        raise Exception("Security check failed: Forbidden functions found in the plugin code.")
+                    self.api.logger.info(f"Security check passed for plugin: {plugin_dir}")
+                    self._load_plugin(plugin_dir, container)
+                    self.installed_plugins[plugin_name] = plugin_dir
         except Exception as e:
             self.api.logger.error(f"Plugin installation failed: {e}")
             # インストールに失敗した場合は展開したフォルダを削除
@@ -135,25 +209,37 @@ class PluginManager:
         self.api.logger.info(unique_key)
 
         def on_rm_error(func, path, exc_info) -> None:
+            if exc_info[0] == FileNotFoundError:
+                return
             import stat
             os.chmod(path, stat.S_IWRITE)
             os.unlink(path)
-        # プラグインディレクトリを削除
+
         shutil.rmtree(plugin_dir, onerror=on_rm_error)
-        # UIからプラグイン関連の要素を削除
+
         if unique_key in self.plugin_dict:
             ui_element = self.plugin_dict.pop(unique_key)
             if ui_element in self.myapp_container.controls:
                 self.myapp_container.controls.remove(ui_element)
             else:
                 self.api.logger.error("no ui element which matches unique_key")
-        # 削除確認ダイアログを閉じる
+
+            self.installed_plugins.pop(unique_key, None)
+        else:
+            self.api.logger.error(f"plugin not found in plugin_dict: {unique_key}")
+
         self.page.dialog.open = False
         self.page.update()
 
     def load_installed_plugins(self, container: ft.Container) -> None:
         self.plugin_dict = {}
         target_list = [filename for filename in os.listdir(self.plugin_folder_path) if not filename.startswith('.')]
+
+        # 存在しないプラグインをself.installed_pluginsから削除する
+        for plugin_name in list(self.installed_plugins.keys()):
+            if plugin_name not in target_list:
+                del self.installed_plugins[plugin_name]
+
         for plugin_name in target_list:
             self.api.logger.info(plugin_name)
             plugin_dir = os.path.join(self.plugin_folder_path, plugin_name)
